devtools/client/shared/components/reps/test/mochitest/test_reps_grip-map.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test GripMap rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - GripMap</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
"use strict";

window.onload = Task.async(function* () {
  const {
    REPS,
    MODE,
    getRep,
    getSelectableInInspectorGrips,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { GripMap } = REPS;

  const componentUnderTest = GripMap;

  try {
    yield testEmptyMap();
    yield testSymbolKeyedMap();
    yield testWeakMap();

    // // Test entries iterator
    yield testMaxEntries();
    yield testMoreThanMaxEntries();
    yield testUninterestingEntries();

    yield testOnDomNodeMouseOver();
    yield testOnDomNodeMouseOut();
    yield testOnDomNodeInspectIconClick();

    yield testObjectLink();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testEmptyMap() {
    // Test object: `new Map()`
    const testName = "testEmptyMap";

    // Test that correct rep is chosen
    const gripStub = getGripStub("testEmptyMap");
    is(getRep(gripStub), GripMap.rep, "Rep correctly selects GripMap Rep");

    // Test rendering
    const defaultOutput = `Map {  }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: "Map",
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testSymbolKeyedMap() {
    // Test object:
    // `new Map([[Symbol("a"), "value-a"], [Symbol("b"), "value-b"]])`
    const testName = "testSymbolKeyedMap";

    const defaultOutput = `Map { Symbol(a): "value-a", Symbol(b): "value-b" }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: "Map",
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testWeakMap() {
    // Test object: `new WeakMap([[{a: "key-a"}, "value-a"]])`
    const testName = "testWeakMap";

    // Test that correct rep is chosen
    const gripStub = getGripStub("testWeakMap");
    is(getRep(gripStub), GripMap.rep, "Rep correctly selects GripMap Rep");

    // Test rendering
    const defaultOutput = `WeakMap { Object: "value-a" }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: "WeakMap",
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      },
      {
        // Check the custom title with nested objects to make sure nested objects are not
        // displayed with their parent's title.
        mode: MODE.LONG,
        title: "CustomTitle",
        expectedOutput: `CustomTitle { Object: "value-a" }`,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testMaxEntries() {
    // Test object:
    // `new Map([["key-a","value-a"], ["key-b","value-b"], ["key-c","value-c"]])`
    const testName = "testMaxEntries";

    const defaultOutput = `Map { "key-a": "value-a", "key-b": "value-b", "key-c": "value-c" }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: "Map",
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testMoreThanMaxEntries() {
    // Test object = `new Map(
    //   [["key-0", "value-0"], ["key-1", "value-1"]], …, ["key-100", "value-100"]]}`
    const testName = "testMoreThanMaxEntries";

    const defaultOutput =
      `Map { "key-0": "value-0", "key-1": "value-1", "key-2": "value-2", 98 more… }`;

    // Generate string with 101 entries, which is the max limit for 'long' mode.
    let longString = Array.from({length: 10}).map((_, i) => `"key-${i}": "value-${i}"`);
    const longOutput = `Map { ${longString.join(", ")}, 91 more… }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Map`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: longOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testUninterestingEntries() {
    // Test object:
    // `new Map([["key-a",null], ["key-b",undefined], ["key-c","value-c"], ["key-d",4]])`
    const testName = "testUninterestingEntries";

    const defaultOutput =
      `Map { "key-a": null, "key-c": "value-c", "key-d": 4, 1 more… }`;
    const longOutput =
      `Map { "key-a": null, "key-b": undefined, "key-c": "value-c", "key-d": 4 }`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `Map`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: longOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testOnDomNodeMouseOver() {
    const nodeValuedStub = getGripStub("testNodeValuedMap");
    const nodeKeyedStub = getGripStub("testNodeKeyedMap");

    const valuesGrips = getSelectableInInspectorGrips(nodeValuedStub);
    is(valuesGrips.length, 3, "the stub has three node grips");

    const keysGrips = getSelectableInInspectorGrips(nodeKeyedStub);
    is(keysGrips.length, 3, "the stub has three node grips");

    let mouseOverValue;
    let onDOMNodeMouseOver = (object) => {
      mouseOverValue = object;
    };

    info("Testing onDOMNodeMouseOver on node valued Map");
    const nodeValuedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeValuedStub,
      onDOMNodeMouseOver,
    });

    let nodes = nodeValuedRenderedComponent.querySelectorAll(".objectBox-node");
    nodes.forEach((node, index) => {
      TestUtils.Simulate.mouseOver(node);
      is(mouseOverValue, valuesGrips[index],
        "onDOMNodeMouseOver is called with the expected argument " +
        "when mouseover is fired on the Rep");
    });

    info("Testing onDOMNodeMouseOver on node keyed Map");
    const nodeKeyedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeKeyedStub,
      onDOMNodeMouseOver,
    });

    nodes = nodeKeyedRenderedComponent.querySelectorAll(".objectBox-node");
    nodes.forEach((node, index) => {
      TestUtils.Simulate.mouseOver(node);
      is(mouseOverValue, keysGrips[index],
        "onDOMNodeMouseOver is called with the expected argument " +
        "when mouseover is fired on the Rep");
    });
  }

  function testOnDomNodeMouseOut() {
    const nodeValuedStub = getGripStub("testNodeValuedMap");
    const nodeKeyedStub = getGripStub("testNodeKeyedMap");

    const valuesGrips = getSelectableInInspectorGrips(nodeValuedStub);
    is(valuesGrips.length, 3, "the stub has three node grips");

    const keysGrips = getSelectableInInspectorGrips(nodeKeyedStub);
    is(keysGrips.length, 3, "the stub has three node grips");

    let called = 0;
    let onDOMNodeMouseOut = (object) => {
      called++;
    };

    info("Testing onDOMNodeMouseOut on node valued Map");
    const nodeValuedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeValuedStub,
      onDOMNodeMouseOut,
    });

    let nodes = nodeValuedRenderedComponent.querySelectorAll(".objectBox-node");
    info("Simulating mouseout on each value node");
    nodes.forEach((node, index) => TestUtils.Simulate.mouseOut(node));
    is(called, 3,
      "onDOMNodeMouseOut is called when mouseout is fired on each value NodeRep");

    info("Testing onDOMNodeMouseOut on node keyed Map");
    const nodeKeyedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeKeyedStub,
      onDOMNodeMouseOut,
    });

    nodes = nodeKeyedRenderedComponent.querySelectorAll(".objectBox-node");
    // Resets counter
    called = 0;
    info("Simulating mouseout on each key node");
    nodes.forEach((node, index) => TestUtils.Simulate.mouseOut(node));
    is(called, 3,
      "onDOMNodeMouseOut is called when mouseout is fired on each key NodeRep");
  }

  function testOnDomNodeInspectIconClick() {
    const nodeValuedStub = getGripStub("testNodeValuedMap");
    const nodeKeyedStub = getGripStub("testNodeKeyedMap");

    const valuesGrips = getSelectableInInspectorGrips(nodeValuedStub);
    is(valuesGrips.length, 3, "the stub has three node grips");

    const keysGrips = getSelectableInInspectorGrips(nodeKeyedStub);
    is(keysGrips.length, 3, "the stub has three node grips");

    let inspectIconClickedValue = null;
    let onInspectIconClick = (object) => {
      inspectIconClickedValue = object;
    };

    const renderedComponentWithoutInspectIcon = renderComponent(GripMap.rep, {
      object: getGripStub("testDisconnectedNodeValuedMap"),
      onInspectIconClick,
    });
    is(renderedComponentWithoutInspectIcon.querySelector(".open-inspector"), null,
      "There isn't an inspect icon when nodes are not connected to the DOM tree");

    info("Testing onInspectIconClick on node valued Map");
    const nodeValuedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeValuedStub,
      onInspectIconClick,
    });

    let icons = nodeValuedRenderedComponent.querySelectorAll(".open-inspector");
    is(icons.length, valuesGrips.length,
      "There is an icon for each node connected to the DOM tree");

    icons.forEach((icon, index) => {
      TestUtils.Simulate.click(icon);
      is(inspectIconClickedValue, valuesGrips[index], "onInspectIconClick is called " +
        "with the expected argument when the inspect icon is clicked");
    });

    info("Testing onInspectIconClick on node keyed Map");
    const nodeKeyedRenderedComponent = renderComponent(GripMap.rep, {
      object: nodeKeyedStub,
      onInspectIconClick,
    });

    icons = nodeKeyedRenderedComponent.querySelectorAll(".open-inspector");
    is(icons.length, keysGrips.length,
      "There is an icon for each node connected to the DOM tree");

    icons.forEach((icon, index) => {
      TestUtils.Simulate.click(icon);
      is(inspectIconClickedValue, keysGrips[index], "onInspectIconClick is called " +
        "with the expected argument when the inspect icon is clicked");
    });
  }

  function testObjectLink() {
    // Test object:
    // `new Map([["key-a",null], ["key-b",undefined], ["key-c","value-c"], ["key-d",4]])`
    const defaultOutput =
      `*Map** { *"key-a": null, "key-c": "value-c", "key-d": 4, *1 more…** }*`;
    const longOutput =
      `*Map** { *"key-a": null, "key-b": undefined, "key-c": "value-c", "key-d": 4* }*`;

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: `*Map*`,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: longOutput,
      }
    ];

    testRepRenderModes(
      modeTests,
      "testObjectLink",
      componentUnderTest,
      getGripStub("testUninterestingEntries"),
      {
        objectLink: (props, ...children) => React.DOM.span({},
          "*", ...children, "*"),
      }
    );
  }

  function getGripStub(functionName) {
    switch (functionName) {
      case "testEmptyMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj97",
          "class": "Map",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 0,
            "entries": []
          }
        };

      case "testSymbolKeyedMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj118",
          "class": "Map",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 2,
            "entries": [
              [
                {
                  "type": "symbol",
                  "name": "a"
                },
                "value-a"
              ],
              [
                {
                  "type": "symbol",
                  "name": "b"
                },
                "value-b"
              ]
            ]
          }
        };

      case "testWeakMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj115",
          "class": "WeakMap",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 1,
            "entries": [
              [
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj116",
                  "class": "Object",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 1
                },
                "value-a"
              ]
            ]
          }
        };

      case "testMaxEntries":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj109",
          "class": "Map",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 3,
            "entries": [
              [
                "key-a",
                "value-a"
              ],
              [
                "key-b",
                "value-b"
              ],
              [
                "key-c",
                "value-c"
              ]
            ]
          }
        };

      case "testMoreThanMaxEntries": {
        let entryNb = 101;
        return {
          "type": "object",
          "class": "Map",
          "actor": "server1.conn0.obj332",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": entryNb,
            // Generate 101 entries, which is more that the maximum
            // limit in case of the 'long' mode.
            "entries": Array.from({length: entryNb}).map((_, i) => {
              return [`key-${i}`, `value-${i}`];
            })
          }
        };
      }

      case "testUninterestingEntries":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj111",
          "class": "Map",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 4,
            "entries": [
              [
                "key-a",
                {
                  "type": "null"
                }
              ],
              [
                "key-b",
                {
                  "type": "undefined"
                }
              ],
              [
                "key-c",
                "value-c"
              ],
              [
                "key-d",
                4
              ]
            ]
          }
        };

      case "testDisconnectedNodeValuedMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj213",
          "class": "Map",
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 3,
            "entries": [
              [
                "item-0",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj214",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": false,
                    "attributes": {
                      "id": "btn-1",
                      "class": "btn btn-log",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ],
              [
                "item-1",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj215",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": false,
                    "attributes": {
                      "id": "btn-2",
                      "class": "btn btn-err",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ],
              [
                "item-2",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj216",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": false,
                    "attributes": {
                      "id": "btn-3",
                      "class": "btn btn-count",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ]
            ]
          }
        };

      case "testNodeValuedMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj213",
          "class": "Map",
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 3,
            "entries": [
              [
                "item-0",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj214",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-1",
                      "class": "btn btn-log",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ],
              [
                "item-1",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj215",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-2",
                      "class": "btn btn-err",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ],
              [
                "item-2",
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj216",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-3",
                      "class": "btn btn-count",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                }
              ]
            ]
          }
        };

      case "testNodeKeyedMap":
        return {
          "type": "object",
          "actor": "server1.conn1.child1/obj223",
          "class": "WeakMap",
          "ownPropertyLength": 0,
          "preview": {
            "kind": "MapLike",
            "size": 3,
            "entries": [
              [
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj224",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-1",
                      "class": "btn btn-log",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                },
                "item-0"
              ],
              [
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj225",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-3",
                      "class": "btn btn-count",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                },
                "item-2"
              ],
              [
                {
                  "type": "object",
                  "actor": "server1.conn1.child1/obj226",
                  "class": "HTMLButtonElement",
                  "extensible": true,
                  "frozen": false,
                  "sealed": false,
                  "ownPropertyLength": 0,
                  "preview": {
                    "kind": "DOMNode",
                    "nodeType": 1,
                    "nodeName": "button",
                    "isConnected": true,
                    "attributes": {
                      "id": "btn-2",
                      "class": "btn btn-err",
                      "type": "button"
                    },
                    "attributesLength": 3
                  }
                },
                "item-1"
              ]
            ]
          }
        };
    }
  }
});
</script>
</pre>
</body>
</html>
